feat(ngrok people counting): add a people counting plugin which exposes an enpoint via ngrok/fastapi
Users can trigger the people counting analysis on a
url pointing to a video file.

This is done with a HTTP GET on the enpoint get_people_count
with the parameter 'url'. This will raise a child pipeline
on the same node which will periodically send the people count.
Once the analysis is finished the child pipeline is closed and
the analysis of the url marked as done and cached for future
replies.